rename setup to composition funcs
